+++ package.json
@@ -93,7 +93,7 @@
     "recharts": "^2.12.0",
     "remark-gfm": "^4.0.0",
     "sharp": "^0.32.6",
-    "stripe": "^14.16.0",
+    "stripe": "^16.7.0",
     "tailwind-merge": "^2.2.1",
     "tailwindcss-animate": "^1.0.7",
     "uploadthing": "^6.13.2",
