+++ server/stripe.ts
@@ -1,6 +1,6 @@
 import Stripe from 'stripe';
 
 export const stripe = new Stripe(process.env.STRIPE_API_KEY as string, {
-  apiVersion: '2023-10-16',
+  apiVersion: '2024-06-20',
   typescript: true,
 });
